Refactor, Tests and add 'dependency-injector' dependency
Define the Service class hierachy #2
