+++ minos/cqrs/__init__.py
@@ -11,7 +11,7 @@ __email__ = "devs@clariteia.com"
 __version__ = "0.0.1"
 
 
-from .service import (
+from .services import (
     CommandService,
     QueryService,
     Service,
