+++ pyproject.toml
@@ -32,6 +32,7 @@ include = [
 [tool.poetry.dependencies]
 python = "^3.9"
 minos-microservice-common = "^0.1.3"
+dependency-injector = "^4.34.0"
 
 [tool.poetry.dev-dependencies]
 black = "^19.10b"
